Copy game play logic from Window into Model
This adds more logic to the model to handle certain game play events
and prepares to remove them from the window class.

Not all of the code ended up moving to the model, the physics
movements of the player moved into the player class itself, with a
way to ensure we still have colission detection by passing in a
colision detector function from the model.