Replace game logic in window with calls to model
This change ultimately removes the game logic from the window and places
it in the model. This drastically reduces the complexity and the amount
of logic that the window posesses since ultimately the window only needs
to know how to draw the stuff onto the screen and process input events.

The model should be the central source of the game logic. Closes #68