Move WORLD_SIZE into World
Also, give it a better name. The name WORLD_SIZE was actually the width of
the map from the origin to one edge in number of blocks. To make it clear
what we're measuring we now have that stated in the name of where this
value is measuring from and how many units it is measuring in. We also
redefine it back in terms of the width of the world in number of blocks
from edge to edge.